chore: optimize html for inspector
- code refactoring
- add title to webviews in selector
